pyproject.toml
[tool.poetry]
name = "scheduleconverter"
version = "0.1.0"
description = "Converts university timetable to an importable calendar file."
authors = ["Xavi <xavisimpson04@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "^2.0.4"
icalendar = "^5.0.7"
pycodestyle = "^2.11.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
recurring-ical-events = "^2.0.2"
six = "^1.16.0"
tomli = "^2.0.1"
x-wr-timezone = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
